updated train.py w sarsalambda update rename of agents for consistency removed unused mcts training code, since mcts isnt trained (was just for debugging)
