+++ agents/__init__.py
@@ -1,9 +1,9 @@
-from .base import Agent, PlayerAgent
+from .base_agent import Agent, PlayerAgent
 from .random_agent import RandAgent
 from .qlearn_agent import QLearnAgent
 from .mcts_agent import MCTSAgent
 from .sarsa_agent import SarsaAgent
-from .SarsaLambdaAgent import SarsaLambdaAgent
+from .sarsa_lambda_agent import SarsaLambdaAgent
 __all__ = [
     "Agent",
     "PlayerAgent",
