+++ agents/mcts_agent.py
@@ -3,7 +3,7 @@ import random
 from wingedsheep.carcassonne.carcassonne_game import CarcassonneGame
 from wingedsheep.carcassonne.objects.actions.action import Action
 
-from .base import Agent
+from .base_agent import Agent
 from MCTS.node import MCTSNode
 
 from tqdm import tqdm
