+++ agents/sarsa_lambda_agent.py
@@ -3,7 +3,7 @@ import os, pickle, random
 from wingedsheep.carcassonne.carcassonne_game import CarcassonneGame
 from wingedsheep.carcassonne.objects.actions.action import Action
 
-from .base import Agent
+from .base_agent import Agent
 
 
 # Author: Alexander Frolov. Sarsa (Lambda) algorithm implementation.
@@ -166,7 +166,7 @@ class SarsaLambdaAgent(Agent):
         self.last_action_key = self._encode_action(chosen_action) # a'
         self.last_score = current_score
 
-        print(f"{self}: {chosen_action}")
+        # print(f"{self}: {chosen_action}")
         return chosen_action
 
     def reset_episode(self):
